Introduce optional URL argument to all tools
Also include the ADR for this decision.
