+++ src/common/errors.ts
@@ -1,0 +1,6 @@
+export class WikiDiscoveryError extends Error {
+	public constructor( message: string ) {
+		super( message );
+		this.name = 'WikiDiscoveryError';
+	}
+}
